requirements.txt
plotly
streamlit
pandas
Pillow